Fix publish.ps1 - first time run

+++ publish.ps1
@@ -5,7 +5,11 @@ $functionApp = "githubstream-$loc-fn"
 $ErrorActionPreference = 'Stop'
 
 # Package and zip the Function App
-Remove-Item './_functionzip' -Recurse -Force
+if (Test-Path './_functionzip') 
+{ 
+    Remove-Item './_functionzip' -Recurse -Force 
+}
+
 New-Item './_functionzip' -ItemType Directory
 dotnet publish .\GitHubStream\ --configuration Release -o './_functionzip'
 Compress-Archive -Path ./_functionzip/* -DestinationPath ./deployfunction.zip -Force
